Changes Listed in Description
I fixed the cursor showing up and disappearing when required. This was done with changing the UserInputMode and whether or not to show the cursor or not. When the MechanicTipWidget is open, or the EndWidget is open the mouse will be visible with no way to remove it until the player presses the button at which point it will disappear. This code considers every case

I add checks to make sure the MechanicTipWidget and the Reset button only work once a level has been opened. It doesn't require the player to finish the level (something I can change if needed)

Made it so the player can press K to toggle the FPS counter in the top left